+++ rs/sns/governance/src/governance/fail_stuck_upgrade_in_progress_tests.rs
@@ -1,10 +1,8 @@
+use super::assorted_governance_tests::{
+    basic_governance_proto, DoNothingLedger, TEST_GOVERNANCE_CANISTER_ID, TEST_ROOT_CANISTER_ID,
+};
 use crate::{
     governance::{
-        // super
-        tests::{
-            basic_governance_proto, DoNothingLedger, TEST_GOVERNANCE_CANISTER_ID,
-            TEST_ROOT_CANISTER_ID,
-        },
         Governance,
         ValidGovernanceProto,
     },
